requirements.txt
pandas==1.3.5
pypiwin32
terminaltables~=3.1.10
wxpython
requests
auto-py-to-exe
pyinstaller==5.13.2
ntplib
cython==3.0.10
pywinstyles
PyYaml